Move edit -> editPage because git is dumb
