+++ apps/modernization-ui/src/apps/page-builder/page/management/editPage/staticelement/AddStaticElement.spec.tsx
