+++ apps/modernization-ui/src/apps/page-builder/page/management/editPage/staticelement/staticelement.module.scss
